Adding argument to control the configuration used
By adding a -c VALUE you will set the configuration of the VLIW.
There are 12 configuration possibles (0, 1, 2, 3, 4, 5, 8, 9, 10, 11, 12
and 13.
Details on these configurations are given in reconfigureVLIW.h and .cpp
files.

+++ includes/dbt/dbtPlateform.h
@@ -46,7 +46,7 @@ public:
 	VexSimulator* vexSimulator;
 	RiscvSimulator* riscvSimulator;
 
-	short vliwInitialConfiguration;
+	int vliwInitialConfiguration;
 	char vliwInitialIssueWidth;
 	char debugLevel;
